feat: hana adapter
