+++ packages/adapter/src/index.ts
@@ -3,11 +3,11 @@ export * from './adapters/hive'
 export * from './adapters/mysql/index'
 export * from './adapters/postgres/index'
 export * from './adapters/presto'
+export * from './adapters/redshift'
 export * from './adapters/sap-bw'
+export * from './adapters/sap-hana'
 export * from './adapters/trino'
-export * from './adapters/sap-hana' // 依赖包 windows 上打包有问题
 export * from './adapters/xmla'
 export * from './base'
 export * from './helpers'
 export * from './types'
-export * from './adapters/redshift'
